出品者かどうかのチェックを実装

+++ app/controllers/items_controller.rb
@@ -1,5 +1,5 @@
 class ItemsController < ApplicationController
-before_action :authenticate_user!, only: [:new, :create]
+before_action :authenticate_user!, only: [:new, :create, :destroy]
 
   def index
     @items = Item.all.order("created_at DESC")
@@ -25,7 +25,9 @@ before_action :authenticate_user!, only: [:new, :create]
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
+    if current_user == item.user
+      item.destroy
+    end
     redirect_to root_path
   end
